requirements.txt
python-telegram-bot==21.6
motor==3.6.0
python-dotenv==1.0.1
recipe-scrapers==15.1.0
openai==1.52.2
httpx==0.27.2
